requirements-dev.txt
# Testing
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Development Tools
ipython==8.20.0
ipdb==0.13.13

